+++ composer.json
@@ -10,7 +10,8 @@
         "guzzlehttp/guzzle": "^7.0.1",
         "laravel/framework": "^8.54",
         "laravel/sanctum": "^2.11",
-        "laravel/tinker": "^2.5"
+        "laravel/tinker": "^2.5",
+        "rtconner/laravel-likeable": "^3.2"
     },
     "require-dev": {
         "blumilksoftware/codestyle": "^0.6.0",
